Refactoring completed with new class hierarchy. ZMQ communication moved to separate class One can't access class variables anymore, all variables are private and only way to acces them are through set/get functions Added zmq dataType enumeration and zmqBuffer class in separate namespace so it can be used throughout project when needed. Added abstract classes to use as interface as single and group classes (scene and tree items) Added toggable average sampling time to scene Enabled vibration marker dependancy on amplitude

+++ src/qcasusceneitem.h
@@ -41,7 +41,7 @@ public:
     double getValue(dataType key);
     bool getState(dataType key);
     bool isConnected();
-    double getAvgSamplingTime();
+    int getAvgSamplingTime();
 
     QRectF boundingRect() const;
